Split heatpump into separate files and removed unused perf function
