+++ pylesa/heat/constants.py
@@ -1,0 +1,4 @@
+"""Constants used throughout the heat sub-package"""
+ASHP = "ASHP"
+GSHP = "GSHP"
+WSHP = "WSHP"
