+++ pylesa/heat/heatpump.py
@@ -1,113 +1,12 @@
-"""Heat pump module
-
-Modelling a heat pump with modelling approaches of
-simple, lorentz, generic regression, and standard test regression
-
-"""
-import math
 import logging
-from pathlib import Path
-import numpy as np
-import matplotlib.pyplot as plt
-from sklearn.linear_model import LinearRegression
-from sklearn.preprocessing import PolynomialFeatures
 
-from ..io import inputs
+from .constants import ASHP, GSHP, WSHP
 from ..environment import weather
 
+from .lorentz import Lorentz
+from .regression import GenericRegression, StandardTestRegression
+
 LOG = logging.getLogger(__name__)
-
-ASHP = "ASHP"
-GSHP = "GSHP"
-WSHP = "WSHP"
-
-plt.style.use('ggplot')
-plt.rcParams.update({'font.size': 22})
-
-def perf(root: Path, subname: str):
-    root = Path(root).resolve()
-    myInputs = inputs.Inputs(root, subname)
-    input_weather = myInputs.weather()
-    inputs_basics = myInputs.heatpump_basics()
-    modelling_approach = inputs_basics['modelling_approach']
-
-    if modelling_approach == 'Simple':
-
-        inputs_simple = myInputs.heatpump_simple()
-        inputs_demands = myInputs.demands()
-
-        myHeatPump = HeatPump(
-            inputs_basics['heat_pump_type'],
-            inputs_basics['modelling_approach'],
-            inputs_basics['capacity'],
-            inputs_basics['ambient_delta_t'],
-            inputs_basics['minimum_runtime'],
-            inputs_basics['minimum_output'],
-            inputs_basics['data_input'],
-            inputs_demands['source_temp'],
-            inputs_demands['return_temp_DH'],
-            input_weather,
-            simple_cop=inputs_simple)
-
-        return myHeatPump.performance()
-
-    elif modelling_approach == 'Lorentz':
-
-        inputs_lorentz = myInputs.heatpump_lorentz()
-        inputs_demands = myInputs.demands()
-
-        myHeatPump = HeatPump(
-            inputs_basics['heat_pump_type'],
-            inputs_basics['modelling_approach'],
-            inputs_basics['capacity'],
-            inputs_basics['ambient_delta_t'],
-            inputs_basics['minimum_runtime'],
-            inputs_basics['minimum_output'],
-            inputs_basics['data_input'],
-            inputs_demands['source_temp'],
-            inputs_demands['return_temp_DH'],
-            input_weather,
-            lorentz_inputs=inputs_lorentz)
-
-        return myHeatPump.performance()
-
-    elif modelling_approach == 'Generic regression':
-
-        inputs_demands = myInputs.demands()
-
-        myHeatPump = HeatPump(
-            inputs_basics['heat_pump_type'],
-            inputs_basics['modelling_approach'],
-            inputs_basics['capacity'],
-            inputs_basics['ambient_delta_t'],
-            inputs_basics['minimum_runtime'],
-            inputs_basics['minimum_output'],
-            inputs_basics['data_input'],
-            inputs_demands['source_temp'],
-            inputs_demands['return_temp_DH'],
-            input_weather)
-
-        return myHeatPump.performance()
-
-    elif modelling_approach == 'Standard test regression':
-
-        inputs_standard_regression = myInputs.heatpump_standard_regression()
-        inputs_demands = myInputs.demands()
-
-        myHeatPump = HeatPump(
-            inputs_basics['heat_pump_type'],
-            inputs_basics['modelling_approach'],
-            inputs_basics['capacity'],
-            inputs_basics['ambient_delta_t'],
-            inputs_basics['minimum_runtime'],
-            inputs_basics['minimum_output'],
-            inputs_basics['data_input'],
-            inputs_demands['source_temp'],
-            inputs_demands['return_temp_DH'],
-            input_weather,
-            standard_test_regression_inputs=inputs_standard_regression)
-
-        return myHeatPump.performance()
 
 
 class HeatPump(object):
@@ -368,251 +267,3 @@ class HeatPump(object):
              'hp_elec': hp_elec}
 
         return d
-
-
-class Lorentz(object):
-
-    def __init__(self, cop, flow_temp_spec, return_temp_spec,
-                 ambient_temp_in_spec, ambient_temp_out_spec,
-                 elec_capacity):
-        """lorentz calculations and attributes
-
-        based on EnergyPRO method
-
-        Arguments:
-            cop {float} -- cop at specified conditions
-            flow_temp_spec {float} -- temperature from HP spec
-            return_temp_spec {float} -- tempature to HP spec
-            ambient_temp_in_spec {float} -- specificed
-            ambient_temp_out_spec {float} -- spec
-            elec_capacity {float} -- absolute
-        """
-
-        self.cop = cop
-        self.flow_temp_spec = flow_temp_spec
-        self.return_temp_spec = return_temp_spec
-        self.ambient_temp_in_spec = ambient_temp_in_spec
-        self.ambient_temp_out_spec = ambient_temp_out_spec
-        self.elec_capacity = elec_capacity
-
-    def hp_eff(self):
-        """heat pump efficiency which is static
-
-        # calcultaions of the lorentz model. starting with the mean temps fo
-        # for the temp flow and return of heat pump, t high mean
-        # and the ambient in and out temps, t low mean
-
-        Returns:
-            float -- efficiency of the heat pump
-        """
-        t_high_mean = ((self.flow_temp_spec - self.return_temp_spec) /
-                       (math.log((self.flow_temp_spec + 273.15) /
-                                 (self.return_temp_spec + 273.15))))
-
-        t_low_mean = (
-            (self.ambient_temp_in_spec - self.ambient_temp_out_spec) /
-            (math.log((self.ambient_temp_in_spec + 273.15) /
-                      (self.ambient_temp_out_spec + 273.15))))
-
-        # lorentz cop is the highest theoretical cop
-        cop_lorentz = t_high_mean / (t_high_mean - t_low_mean)
-
-        # this gives the heat pump efficiency using the stated cop
-        # the lorentz cop is calcualted for each timestep
-        # then this is multiplied by the heat pump
-        # efficiency to give actual cop
-        hp_eff = self.cop / cop_lorentz
-
-        return hp_eff
-
-    def calc_cop(self, hp_eff, flow_temp, return_temp,
-                 ambient_temp_in, ambient_temp_out):
-        """cop for timestep
-
-        calculates the cop based upon actual flow/retur and ambient
-        uses heat pump efficiency from before
-
-        Arguments:
-            hp_eff {float} -- heat pump efficiency
-            flow_temp {float} -- flow temperature from heat pump
-            return_temp {float} -- temperature returning to heat pump
-            ambient_temp_in {float} -- real-time
-            ambient_temp_out {float} -- real-time
-
-        Returns:
-            float -- cop for timestep
-        """
-
-        t_high_mean = ((flow_temp - return_temp) /
-                       (math.log((flow_temp + 273.15) /
-                                 (return_temp + 273.15))))
-
-        t_low_mean = ((ambient_temp_in - ambient_temp_out) /
-                      (math.log((ambient_temp_in + 273.15) /
-                                (ambient_temp_out + 273.15))))
-
-        cop_lorentz = t_high_mean / (t_high_mean - t_low_mean)
-
-        cop = hp_eff * cop_lorentz
-
-        return cop
-
-    def calc_duty(self, capacity):
-        """duty for timestep
-
-        calculates duty for timestep, ensures this is not exceeded
-
-        Arguments:
-            capacity {float} -- electrical capacity of heat pump
-
-        Returns:
-            float -- duty is the thermal output of the heat pump
-        """
-
-        duty_max = self.cop * self.elec_capacity
-        if duty_max >= capacity:
-            duty = capacity
-        elif duty_max < capacity:
-            duty = duty_max
-
-        return duty
-
-
-class GenericRegression(object):
-    """uses generic regression analysis to predict performance
-
-    see Staffel paper on review of domestic heat pumps for coefficients
-    """
-
-    def ASHP_cop(self, flow_temp, ambient_temp):
-
-        cop = (6.81 -
-               0.121 * (flow_temp - ambient_temp) +
-               0.00063 * (flow_temp - ambient_temp) ** 2
-               )
-        return cop
-
-    def ASHP_duty(self, flow_temp, ambient_temp):
-
-        duty = (5.80 +
-                0.21 * (ambient_temp)
-                )
-        return duty
-
-    def GSHP_cop(self, flow_temp, ambient_temp):
-
-        cop = (8.77 -
-               0.15 * (flow_temp - ambient_temp) +
-               0.000734 * (flow_temp - ambient_temp) ** 2
-               )
-        return cop
-
-    def GSHP_duty(self, flow_temp, ambient_temp):
-
-        duty = (9.37 +
-                0.30 * ambient_temp
-                )
-        return duty
-
-    def plot_cop(self):
-
-        x_ambient = np.linspace(-5, 15, num=100)
-
-        cop_55 = []
-        for z in range(len(x_ambient)):
-            c = self.ASHP_cop(55, x_ambient[z])
-            if x_ambient[z] <= 5:
-                c = 0.9 * c
-            cop_55.append(c)
-
-        cop_45 = []
-        for z in range(len(x_ambient)):
-            c = self.ASHP_cop(45, x_ambient[z])
-            if x_ambient[z] <= 5:
-                c = 0.9 * c
-            cop_45.append(c)
-
-        plt.plot(x_ambient, cop_45, linewidth=2)
-        plt.plot(x_ambient, cop_55, linewidth=2)
-        plt.legend(['Flow T 45', 'FLow T 55'], loc='best')
-        plt.ylabel('COP')
-        plt.xlabel('Ambient temperature')
-        plt.show()
-
-
-class StandardTestRegression(object):
-
-    def __init__(self, data_x, data_COSP,
-                 data_duty, degree=2):
-        """regression analysis based on standard test condition data
-
-        trains a model
-        predicts cop and duty
-
-        Arguments:
-            data_x {dataframe} -- with flow temps and ambient temps
-            data_COSP {dataframe} -- with cosp data for different data_X
-            data_duty {dataframe} -- with duty data for different data_X
-
-        Keyword Arguments:
-            degree {number} -- polynomial number (default: {2})
-        """
-        self.data_x = data_x
-        self.data_COSP = data_COSP
-        self.data_duty = data_duty
-        self.degree = degree
-
-    def train(self):
-        """training model
-        """
-        poly = PolynomialFeatures(degree=self.degree, include_bias=False)
-
-        X_new = poly.fit_transform(self.data_x)
-        Y_COSP_new = poly.fit_transform(self.data_COSP)
-        Y_duty_new = poly.fit_transform(self.data_duty)
-
-        model_cop = LinearRegression()
-        model_cop.fit(X_new, Y_COSP_new)
-
-        model_duty = LinearRegression()
-        model_duty.fit(X_new, Y_duty_new)
-
-        return {'COP_model': model_cop, 'duty_model': model_duty}
-
-    def predict_COP(self, model, ambient_temp, flow_temp):
-        """predicts COP from model
-
-        Arguments:
-            model {dic} -- cop and duty models in dic
-            ambient_temp {float} --
-            flow_temp {float} --
-
-        Returns:
-            float -- predicted COP
-        """
-
-        x_pred = np.array([ambient_temp, flow_temp]).reshape(1, -1)
-        poly = PolynomialFeatures(degree=2, include_bias=False)
-        x_pred_new = poly.fit_transform(x_pred)
-        pred_cop = model.predict(x_pred_new)
-
-        return float(pred_cop[:, 0])
-
-    def predict_duty(self, model, ambient_temp, flow_temp):
-        """predicts duty from regression model
-
-        Arguments:
-            model {dic} -- cop and duty models in dic
-            ambient_temp {float} --
-            flow_temp {float} --
-
-        Returns:
-            float -- predicted COP
-        """
-
-        x_pred = np.array([ambient_temp, flow_temp]).reshape(1, -1)
-        poly = PolynomialFeatures(degree=2, include_bias=False)
-        x_pred_new = poly.fit_transform(x_pred)
-        pred_duty = model.predict(x_pred_new)
-
-        return float(pred_duty[:, 0])
